+++ modulo5/introducao-autenticacao/package.json
@@ -19,10 +19,12 @@
   },
   "dependencies": {
     "@types/knex": "^0.16.1",
+    "@types/uuid": "^8.3.4",
     "cors": "^2.8.5",
     "dotenv": "^16.0.2",
     "express": "^4.18.1",
     "knex": "^2.3.0",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "uuid": "^9.0.0"
   }
 }
