+++ modulo5/introducao-autenticacao/services/generateId.ts
@@ -1,0 +1,5 @@
+import { v4 } from "uuid";
+
+export function generateId(): string {
+    return v4();
+  }
