+++ modulo5/introducao-autenticacao/src/index.ts
@@ -1,0 +1,44 @@
+import express, { Express } from "express";
+import knex from "knex";
+import cors from "cors";
+import dotenv from "dotenv";
+import { AddressInfo } from "net";
+
+
+dotenv.config();
+
+export const connection = knex({
+	client: "mysql",
+	connection: {
+    host: process.env.DB_HOST,
+    port: 3306,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASS,
+    database: process.env.DB_NAME
+  }
+});
+
+const app: Express = express();
+app.use(express.json());
+app.use(cors());
+
+const userTableName = "Usuariointro";
+const createUser = async (id: string, email: string, password: string) => {
+    await connection
+      .insert({
+        id,
+        email,
+        password,
+      })
+     .into(userTableName);
+  };
+
+
+const server = app.listen(process.env.PORT || 3003, () => {
+    if (server) {
+       const address = server.address() as AddressInfo;
+       console.log(`Server is running in http://localhost: ${address.port}`);
+    } else {
+       console.error(`Failure upon starting server.`);
+    }
+});
