+++ modulo5/introducao-autenticacao/src/table.sql
@@ -1,0 +1,6 @@
+-- Active: 1662330059941@@35.226.146.116@3306@alves-bruna-sousa
+CREATE TABLE Usuariointro (
+		id VARCHAR(255) PRIMARY KEY,
+    email VARCHAR(255) UNIQUE NOT NULL,
+    password VARCHAR(255) NOT NULL
+);
